Refactoring of types to ensure that we separate types used in API from types that are not visible from API
