+++ zenoh-protocol/src/session/mod.rs
@@ -24,7 +24,8 @@ use async_std::sync::{Arc, Weak};
 use async_trait::async_trait;
 
 use crate::link::Link;
-use crate::proto::{SessionMessage, WhatAmI, ZenohMessage};
+use crate::core::WhatAmI;
+use crate::proto::{SessionMessage, ZenohMessage};
 
 use zenoh_util::{zerror, zweak};
 use zenoh_util::core::{ZError, ZErrorKind, ZResult};
