+++ zenoh-router/src/runtime/mod.rs
@@ -16,9 +16,8 @@ use std::time::Duration;
 use async_std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
 use zenoh_util::core::{ZResult, ZError, ZErrorKind};
 use zenoh_util::zerror;
-use zenoh_protocol::core::PeerId;
+use zenoh_protocol::core::{PeerId, WhatAmI, whatami};
 use zenoh_protocol::link::Locator;
-use zenoh_protocol::proto::{WhatAmI, whatami};
 use zenoh_protocol::session::{SessionManager, SessionManagerConfig, SessionManagerOptionalConfig};
 use crate::routing::broker::Broker;
 use crate::runtime::orchestrator::SessionOrchestrator;
