+++ zenoh-router/src/runtime/orchestrator.rs
@@ -18,7 +18,8 @@ use socket2::{Socket, Domain, Type};
 use zenoh_util::core::{ZResult, ZError, ZErrorKind};
 use zenoh_util::zerror;
 use zenoh_protocol::io::{WBuf, RBuf};
-use zenoh_protocol::proto::{WhatAmI, whatami, SessionMessage, SessionBody};
+use zenoh_protocol::core::{WhatAmI, whatami};
+use zenoh_protocol::proto::{SessionMessage, SessionBody};
 use zenoh_protocol::link::Locator;
 use zenoh_protocol::session::SessionManager;
 use crate::runtime::Config;
